Fix .srodata and .sdata2 in linker scripts
.srodata should really be in the .text section. .sdata2 should be
in the .text section too, as GCC 9 seems to put some variables in
that section instead of .srodata.

This fixes an issue as seen with image built by GCC 10.1.0, that
globalInitFunctions and spanOfGlobalInitFunctions variables are
put in .srodata which ends up being in the L2LIM address range,
but at that time L2LIM contains all zeroes, causing boot failures.

Signed-off-by: Bin Meng <bin.meng@windriver.com>