Explicitly specify the compiler flags for RISC-V arch & abi
We should not depend on the toolchain default settings for the RISC-V
arch and abi. Instead we should explicitly specify them.

Current makefiles use an undocumented CONFIG_WITH_ARCH shell variable
to pass arch & abi during build, but this should not be a user choice.

This resolves the link error as seen with the bootlin toolchain.

Fixes: https://github.com/polarfire-soc/hart-software-services/issues/19
Signed-off-by: Bin Meng <bmeng.cn@gmail.com>