(chore)debug: Update debug configurations
Updated the included debug configuration name to come in line with bare
metal examples, it is now post fixed with "hw all-harts debug" to
indicate it is a download and debug configuration.

Added a "hw all-harts attach" debug configuration to attach to a running
HSS from SoftConsole and only load symbols for the application and not
set the program counters to _start.