fix(version): provide a fallback version for non-git downloads
Non-git downloads, eg the tarballs from GitHub, complain about fatal
errors for each file that is compiled. The build still succeeds, but the
output looks very noisy.
The source of the complaints is a call to git that sets the version
number of the program. Provide a fallback to "unknown" in the cases
where the command fails and silence the failure.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>