fix(reboot): assert if a warm reboot is requested without SERVICE_BOOT
The use case for running without SERVICE_BOOT is when "you want one to
setup the platform/DDR training, but then not run any more code.
Its kinda unusual, but it is handy."

Any reboot requested in this context will be from the HSS CLI, which are
always cold reboots.
There is no provider for warm reboots without SERVICE_BOOT, so if a warm
reboot has been requested, inform the user and drop them into a
debugger.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>