fix(tinycli): Only compile boot related code when SERVICE_BOOT is set
Update the tinyCLI service to allow disabling the CONFIG_SERVICE_BOOT.

After disabling SERVICE_BOOT in menuconfig, pBootImage was highlighted
as being an undefined reference.There appear to be a bunch of other
bits inside tinycli that should also be disabled if SERVICE_BOOT is
turned off.

By disabling CONFIG_SERVICE_BOOT, an eNVM HSS image can
be created that trains DDR, configures the mpfs-hal (and
L2 Scratchpad), but subsequently idles or drops to the tinyCLI.

This facilitates rebuilding the HSS and interactive debugging
of it.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>