chore(reboot): unify error reporting for auto-upgrade
While the authentication service is required to check whether an upgrade
image is valid (otherwise the system may lock up), it does not fail when
the upgrade image is identical to that currently programmed. Instead, in
this case, the failure to upgrade is observed while running the auto
upgrade server.
Unify the error handling for both services called by
HSS_reboot_auto_upgrade() so that this is reported to the user, as
intended.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>