fix(reboot): ensure reboot printout appears before Auto Update
The print that is supposed to inform the user that IAP is in progress
appears *after* the service is requested, thus races against the service
commencing.
Move it before the service request & add a delay to make sure the uart
hardware gets the characters out before the MSS is frozen.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>