services: remove unclear ecall return value dance
Our vendor ecalls seem to all contain a dance like:
	if (result >= 0) {
		*out_val = result;
		result = 0;
	}

For some context, `result` here eventually becomes the value returned in
a0 to the "ecall-er" and `out_val` is the value returned in a1.

In this dance, if the result of the ecall's implementation is not an
error, `out_val` is set to `result` and `result` is cleared. The purpose
of this code is not clear without investigating the code further, nor is
it immediately apparent that assigning zero to result is actually an
assignment of SBI_OK.

In the vast majority of cases, this dance is a waste of time as we do
not have positive, non-zero values in `result`, and out_val is zero
initialised in sbi_ecall_handler(), before the extensions handler is
even called.

There is currently one case where we do actually want to return
something to the caller of the ecall in a1, where the dance actually
does something. However, the code's intent would be much clearer if
the assignment to out_val was done directly.

The other upside to this change is that assignments to `result` can be
reduced to only the standard SBI ecall return codes.

Remove the dance from our vendor ecalls and set `result` only to the
set of standard SBI ecall return codes. Convert the only positive,
non-zero (ab)use of `result` to a direct assignment to `out_val`.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>