feat(ddr): run DDR training as service
This refactoring allows the DDR training to run as a service.
This has the added benefit of ensuring that early DDR training is
guarded by the watchdog.

Furthermore, by adding a generic triggering mechanism, it allows
us to switch DDR training to U54_1, thereby running it in parallel
with other activities on the E51.

Signed-off-by: Ivan Griffin <ivan.griffin@microchip.com>