fix(atomic): remove unnecessary use of volatile
Not sure why volatile is used here.

In C, the volatile keyword is not sufficient to ensure atomicity of accesses
or prevent reordering of instructions by the compiler.

The volatile keyword is used to indicate to the compiler that a variable's
value may be changed by external means, and as a result the compiler should
not optimize or cache the variable. As a consequence, volatile disables
certain optimizations.

volatile does not provide atomicity guarantees or prevent reordering of
instructions.  For atomic operations, you typically need to use specific
atomic types and functions provided by the platform/library.

In fact, using volatile for atomic synchronization might be misleading and
could lead to incorrect assumptions. While volatile prevents certain
compiler optimizations related to variable access, it does not provide the
necessary atomicity guarantees and memory ordering required for correct
multithreaded programming.

See also
https://www.kernel.org/doc/html/next/process/volatile-considered-harmful.html

Signed-off-by: Ivan Griffin <ivan.griffin@microchip.com>