fix(usbdmsc): tickle watchdog during QSPI flush
The flushing of cached QSPI data can take a long time,
and with the restructuring of initialization into services,
the E51 watchdog is enabled much earlier during startup.

As a result, we need to tickle the watchdog during this
flushing to prevent spurious resets.

We also should synchronize the starting of USBDMSC with
DDR training complete/startup complete.

Signed-off-by: Ivan Griffin <ivan.griffin@microchip.com>