feat(dtb): Add a common DTB for U-Boot to use
Adding a common bare-bones DTB for U-Boot purposes. This
simplifies early boot, allowing a single image to be used across
multiple MPFS systems.

If an application workload wants to provide U-Boot with a custom
non-common DTB, it can do so using either u-boot-dtb.bin (which
ignores the value passed in a1) or it can use ancilliary-data
in the payload generator YAML configuration.

Signed-off-by: Ivan Griffin <ivan.griffin@microchip.com>