ihc: add support for the the Mi-V IHC IP driver v2
Starting from v2024.06 release and onwards, the Icicle Kit reference
design will use the Mi-V IHC IP available in the Libero catalog.

Update driver in the HSS to support the Mi-V IHC IP. THe updates include
changes in the following components:

1. Update MI-V IHC driver to support the mi-v ihc IP driver v2.3

2. Update the remoteproc ecall service to use the Mi-V IHC IP available
in the Libero IP catalog.

3. Add Inter-Processor Communication (IPC) ecall support to the HSS.
The IPC Function IDs (FIDs) provided in the Microchip SBI ECALL
extension define an SBI interface that can be used to send an receive
data between processors or cores in a system.

This IPC SBI interface is compatible with the Mi-V Inter-hart
Communication (IHC) IP available in the Libero IP catalog.

The interface specification is available in the following document:
https://github.com/linux4microchip/microchip-sbi-ecall-extension

4.Update the HSS_USE_IHC Kconfig option to allow signaling E51/u54
during the HSS initialization using the Mi-V IHC IP.

The Mi-V IHC IP version 2 is not backwards compatible with the
Mi-V IHC subsystem used in the Icicle Kit reference design 2024.02
or earlier. For this reason the AMP support in the v2024.06
release will not be backwards compatible with previous releases.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>