fix(uart): Fix issue with UART_SURRENDER
When CONFIG_UART_SURRENDER was enabled, a state transition within
TinyCLI was incorrectly being overwritten, causing it to contend for
UART characters along with application payloads.

The solution is to restruct the code slightly to ensure that
TinyCLI correctly goes into a UART Surrender parking state once boot
has occurred.

Signed-off-by: Ivan Griffin <ivan.griffin@microchip.com>