ihc: add support for the the Mi-V IHC IP driver v2
Starting from the 2025.07 release and onwards,
the Icicle Kit reference design will use the Mi-V IHC IP available in
the Libero catalog.

Add support for the Mi-V IHC v2 driver in the HSS. The updates include
changes in the following components:

1. Add the Mi-V IHC v2 bare-metal driver. Retain the legacy Mi-V IHC
driver to allow customers to continue using it while transitioning.
The Mi-V IHC v2 driver can be enabled using CONFIG_USE_IHC_V2, while
the legacy Mi-V IHC can be enabled using CONFIG_USE_IHC.

2. Add Inter-Processor Communication (IPC) ecall support to the HSS.
The IPC Function IDs (FIDs) provided in the Microchip SBI ecall
extension define an SBI interface that can be used to send and receive
data between processors or cores in a system.

This IPC SBI interface is compatible with the Mi-V Inter-Hart
Communication (IHC) v2 IP available in the Libero IP catalog.

3. Add the remoteproc IPC ecall service to use the Mi-V IHC v2 IP from
the Libero IP catalog.

The interface specification is available in the following document:
https://github.com/linux4microchip/microchip-sbi-ecall-extension

The Mi-V IHC IP version 2 is not backwards compatible with the
Mi-V IHC subsystem used in the Icicle Kit reference design 2025.03
or earlier. For this reason, the AMP support in the v2025.07 release
will not be backwards compatible with previous releases.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>