Update Firmware Transfer Service via YMODEM/JTAG Interface (#1)
# Description

For firmware transfer on NAVC and DPU board, this function is added because there is no USB OTG interface.
To overcome that limitation, YMODEM and JTAG interface is used for transferring binary from host pc to target board.
Also, compression and decompression method are added to reduced transfer binary size.

## Type of change

- [O ] New feature (non-breaking change which adds functionality)